Fix the build images issues
This patch will now properly detect to build the image using OVN rpms
or using OVN/OVS sources.

When rpms to be used, the user should copy all the required rpms - openvswitch
and ovn to this repo directory.

Also moved the cinc Dockerfile to the fedora/cinc folder as we have only support
for fedora images as of now.

Signed-off-by: Numan Siddique <numans@ovn.org>