ovn_cluster.sh start: configure vms using dhcp
This change creates some fake vm ports that use dhcp client in order
to get an ip from ovn. In order to do that, we need a tweak in dhcpc
script to neuter hostname. That tweak is done during container image
creation.

Also enhanced Vagrant provisioning to enable vm nesting. That is now
available on OSX systems with Virtualbox 6.1 or newer.