ovn_cluster.sh: require super powers
As described in
https://github.com/ovn-org/ovn-fake-multinode/issues/24
"sudo docker" instead of "docker" makes scaling tests run a
lot slower. This can cause wasted time when scaling to big
number of interfaces.

Since we already documented uses of this script to be done
by root, this patch removes the sudo prefix in RUNC_CMD and
adds a check right at the top to ensure that the effective
user is root.

Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
Closes ovn-org/ovn-fake-multinode#24