Support IPv6 underlay.
This patch provides a config option - IPV6_UNDERLAY. If it is
set to "yes", then this patch configures the ovsdb-servers to
listen on IPv6 addresses. All the ovn-controllers and ovn-northd
talk to DB servers using IPv6.

The link local address of interface - eth1 is used for this purpose.
I don't see a reason to assign a global IPv6 addresses to each of
the fake node containers.

Signed-off-by: Numan Siddique <numans@ovn.org>