ovn_cluster/count-containers: remove whitespace from container name
when validating the container status on function count-containers
if the name of the container contains any whitespace the container will not be found
by count-containers, and that causes the functions that use count-containers
to run forever and waiting for the desired container to become up.

so we must remove the extra whitespace from the container name if exist.

Signed-off-by: Mohammad Heib <mheib@redhat.com>
Signed-off-by: Mohammad Heib <goody698@gmail.com>