Refactor the install_pkg script
Unify the packages that needs to be installed,
the Fedora and el8 should use the same packages
even though the version might not be the same.

Remove the CentOS repos as they are not
functional, but the whole multinode for ub8
still works.

Last part of this refactor is to getting
custom compilation of iputils, which is
not required anymore and the iputils from
rpm will suffice.

Signed-off-by: Ales Musil <amusil@redhat.com>