Add ability to use ubuntu base images.
This change adds ability to use ubuntu base image by introducing
platform-specific install scripts that are conditionally loaded
during the `cinc` image build. The install script is chosen based
in the value of `$OS_BASE` env. variable.

The change is backward compatible and maintains the default behavior
of building on top of `fedora:latest` image.