ovn_cluster: Make sure we perform cleanup after failure.
Make sure we clear the ovs and ovn directory when the container
buikd fails or is interrupted. This should prevent an issue
when we wouldn't load newer source files because the ovn/ovs
directory was defined. Also remove the handling of tst.rpm and
tst.deb. Those files were not used for anything.

Signed-off-by: Ales Musil <amusil@redhat.com>