vagrant: Switch to CentOS Stream 10 and update provisioning
Update the Vagrant environment to use CentOS Stream 10 base
images instead of Rocky Linux 8. This required adjustments
to repository configuration and package installation.

Also reduced the number of uperf threads in the traffic
script to avoid excessive resource usage.

Note: Adding a GitHub Actions CI test was not possible as
nested VMs are not supported on the default runners.

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>